Refactoring: extracting processors code into separate file for isolating tests
