+++ x-pack/functionbeat/function/beater/proccessors_test.go
@@ -1,0 +1,1 @@
+package beater
